[Trunk] in wbQuickShowConflicts mode, if there are only 2 versions of a record (after applying modgroups), consider this an override without further checking
On the one hand, this means that ITMs or Benign changes are not correctly detected in this mode. 

On the other hand, it can be MUCH faster. e.g. running a -quickshowconflicts on my current TES4 test-setup (all ULs+OCR) went from 11:04 to 1:18 (without a chance in the reported conflict losers).