[Experimental] Adds support for NVSE coSave to FNVDump -Saves.
I am concerned of the way I extract the list of plugins from the coSave. Comments welcome :)

Adds PermanentName to write info on disk, particulary formID, in a way that would resist master adding/sorting and/or changes (for ExportImportText scripts) 
My ultimate goal is to allow a modder to store its plugin in a source version control in a more meaningful way than a single file.

Synced to r1741